Address review comments on PR
