+++ cobalt/android/apk/apk_sources.gni
@@ -18,11 +18,11 @@
 apk_sources = [
   "//starboard/android/apk/app/src/app/java/dev/cobalt/app/CobaltApplication.java",
   "//starboard/android/apk/app/src/app/java/dev/cobalt/app/MainActivity.java",
-  "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/AOSPStarboardBridge.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/ArtworkDownloader.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/ArtworkDownloaderDefault.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/ArtworkLoader.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/AudioPermissionRequester.java",
+  "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/BaseStarboardBridge.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/CaptionSettings.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/CobaltHttpHelper.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/CobaltMediaSession.java",
@@ -35,6 +35,7 @@ apk_sources = [
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/NullCobaltFactory.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/PlatformError.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/ResourceOverlay.java",
+  "//starboard/android/apk/app/src/main/java/dev/cobalt/coat/StarboardBridge.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/libraries/services/clientloginfo/ClientLogInfo.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/libraries/services/clientloginfo/ClientLogInfoModule.java",
   "//starboard/android/apk/app/src/main/java/dev/cobalt/media/AudioOutputManager.java",
