+++ cobalt/android/apk/app/src/app/java/dev/cobalt/app/CobaltApplication.java
@@ -16,7 +16,9 @@ package dev.cobalt.app;
 
 import android.app.Application;
 import android.content.Context;
+
 import dev.cobalt.coat.StarboardBridge;
+
 import org.chromium.base.ApplicationStatus;
 import org.chromium.base.ContextUtils;
 import org.chromium.base.PathUtils;
@@ -24,24 +26,23 @@ import org.chromium.base.library_loader.LibraryLoader;
 import org.chromium.base.library_loader.LibraryProcessType;
 
 /** Android Application hosting the Starboard application. */
-public class CobaltApplication extends Application
-    implements StarboardBridge.HostApplication {
-  private static final String PRIVATE_DATA_DIRECTORY_SUFFIX = "content_shell";
+public class CobaltApplication extends Application implements StarboardBridge.HostApplication {
+    private static final String PRIVATE_DATA_DIRECTORY_SUFFIX = "content_shell";
 
-  StarboardBridge mStarboardBridge;
+    StarboardBridge mStarboardBridge;
 
-  @Override
-  public void setStarboardBridge(StarboardBridge starboardBridge) {
-    mStarboardBridge = starboardBridge;
-  }
+    @Override
+    public void setStarboardBridge(StarboardBridge starboardBridge) {
+        mStarboardBridge = starboardBridge;
+    }
 
-  @Override
-  public StarboardBridge getStarboardBridge() {
-    return mStarboardBridge;
-  }
+    @Override
+    public StarboardBridge getStarboardBridge() {
+        return mStarboardBridge;
+    }
 
-  @Override
-  protected void attachBaseContext(Context base) {
+    @Override
+    protected void attachBaseContext(Context base) {
       super.attachBaseContext(base);
       boolean isBrowserProcess = !ContextUtils.getProcessName().contains(":");
       ContextUtils.initApplicationContext(this);
