+++ starboard/android/shared/base_starboard_bridge.cc
@@ -1,4 +1,4 @@
-// Copyright 2024 The Cobalt Authors. All Rights Reserved.
+// Copyright 2026 The Cobalt Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "starboard/android/shared/starboard_bridge.h"
-
 #include "base/android/jni_array.h"
 #include "base/android/jni_string.h"
 #include "base/strings/string_number_conversions.h"
@@ -21,6 +19,7 @@
 #include "starboard/android/shared/application_android.h"
 #include "starboard/android/shared/file_internal.h"
 #include "starboard/android/shared/log_internal.h"
+#include "starboard/android/shared/starboard_bridge.h"
 #include "starboard/common/command_line.h"
 #include "starboard/common/log.h"
 #include "starboard/common/time.h"
@@ -34,7 +33,7 @@
 #endif  // !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
 
 // Must come after all headers that specialize FromJniType() / ToJniType().
-#include "cobalt/android/jni_headers/AOSPStarboardBridge_jni.h"
+#include "cobalt/android/jni_headers/BaseStarboardBridge_jni.h"
 
 namespace starboard {
 
@@ -66,7 +65,7 @@ std::vector<std::string> GetArgs() {
 
 }  // namespace
 
-jboolean JNI_AOSPStarboardBridge_InitJNI(
+jboolean JNI_BaseStarboardBridge_InitJNI(
     JNIEnv* env,
     const JavaParamRef<jobject>& j_starboard_bridge) {
   SB_CHECK(env);
@@ -78,11 +77,11 @@ jboolean JNI_AOSPStarboardBridge_InitJNI(
   return true;
 }
 
-jlong JNI_AOSPStarboardBridge_CurrentMonotonicTime(JNIEnv* env) {
+jlong JNI_BaseStarboardBridge_CurrentMonotonicTime(JNIEnv* env) {
   return CurrentMonotonicTime();
 }
 
-jlong JNI_AOSPStarboardBridge_StartNativeStarboard(
+jlong JNI_BaseStarboardBridge_StartNativeStarboard(
     JNIEnv* env,
     const JavaParamRef<jobject>& j_asset_manager,
     const JavaParamRef<jstring>& j_files_dir,
@@ -103,22 +102,22 @@ jlong JNI_AOSPStarboardBridge_StartNativeStarboard(
   return reinterpret_cast<jlong>(g_native_app_instance);
 }
 
-void JNI_AOSPStarboardBridge_CloseNativeStarboard(JNIEnv* env, jlong nativeApp) {
+void JNI_BaseStarboardBridge_CloseNativeStarboard(JNIEnv* env,
+                                                  jlong nativeApp) {
   auto* app = reinterpret_cast<ApplicationAndroid*>(nativeApp);
   delete app;
 }
 
-void JNI_AOSPStarboardBridge_InitializePlatformAudioSink(JNIEnv* env) {
+void JNI_BaseStarboardBridge_InitializePlatformAudioSink(JNIEnv* env) {
   SbAudioSinkImpl::Initialize();
 }
 
-void JNI_AOSPStarboardBridge_HandleDeepLink(JNIEnv* env,
-                                        const JavaParamRef<jstring>& jurl,
-                                        jboolean applicationStarted) {
+void JNI_BaseStarboardBridge_HandleDeepLink(JNIEnv* env,
+                                            const JavaParamRef<jstring>& jurl,
+                                            jboolean applicationStarted) {
   // TODO(b/492704919): enable on AOSP when the layering violation is fixed.
 #if !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
   const std::string& url = ConvertJavaStringToUTF8(env, jurl);
-  LOG(INFO) << "StarboardBridge handling DeepLink: " << url;
 
   auto* manager = cobalt::browser::DeepLinkManager::GetInstance();
   if (applicationStarted) {
@@ -131,8 +130,8 @@ void JNI_AOSPStarboardBridge_HandleDeepLink(JNIEnv* env,
 #endif  // !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
 }
 
-void JNI_AOSPStarboardBridge_SetAndroidOSExperience(JNIEnv* env,
-                                                jboolean isAmatiDevice) {
+void JNI_BaseStarboardBridge_SetAndroidOSExperience(JNIEnv* env,
+                                                    jboolean isAmatiDevice) {
   // TODO(b/492704919): enable on AOSP when the layering violation is fixed.
 #if !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
   std::string value = isAmatiDevice ? "Amati" : "Watson";
@@ -143,8 +142,8 @@ void JNI_AOSPStarboardBridge_SetAndroidOSExperience(JNIEnv* env,
 #endif  // !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
 }
 
-void JNI_AOSPStarboardBridge_SetAndroidPlayServicesVersion(JNIEnv* env,
-                                                       jlong version) {
+void JNI_BaseStarboardBridge_SetAndroidPlayServicesVersion(JNIEnv* env,
+                                                           jlong version) {
   // TODO(b/492704919): enable on AOSP when the layering violation is fixed.
 #if !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
   auto header_value_provider =
@@ -155,7 +154,7 @@ void JNI_AOSPStarboardBridge_SetAndroidPlayServicesVersion(JNIEnv* env,
 #endif  // !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
 }
 
-void JNI_AOSPStarboardBridge_SetAndroidBuildFingerprint(
+void JNI_BaseStarboardBridge_SetAndroidBuildFingerprint(
     JNIEnv* env,
     const JavaParamRef<jstring>& fingerprint) {
   // TODO(b/492704919): enable on AOSP when the layering violation is fixed.
@@ -168,7 +167,7 @@ void JNI_AOSPStarboardBridge_SetAndroidBuildFingerprint(
 #endif  // !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
 }
 
-void JNI_AOSPStarboardBridge_SetYoutubeCertificationScope(
+void JNI_BaseStarboardBridge_SetYoutubeCertificationScope(
     JNIEnv* env,
     const JavaParamRef<jstring>& certScope) {
   // TODO(b/492704919): enable on AOSP when the layering violation is fixed.
@@ -181,7 +180,7 @@ void JNI_AOSPStarboardBridge_SetYoutubeCertificationScope(
 #endif  // !BUILDFLAG(IS_PARTNER_TOOLCHAIN)
 }
 
-jboolean JNI_AOSPStarboardBridge_IsReleaseBuild(JNIEnv* env) {
+jboolean JNI_BaseStarboardBridge_IsReleaseBuild(JNIEnv* env) {
 #if BUILDFLAG(COBALT_IS_RELEASE_BUILD)
   return true;
 #else
@@ -189,7 +188,7 @@ jboolean JNI_AOSPStarboardBridge_IsReleaseBuild(JNIEnv* env) {
 #endif
 }
 
-jboolean JNI_AOSPStarboardBridge_IsDevelopmentBuild(JNIEnv* env) {
+jboolean JNI_BaseStarboardBridge_IsDevelopmentBuild(JNIEnv* env) {
 // OFFICIAL_BUILD is set for Cobalt QA and Gold releases
 #if defined(OFFICIAL_BUILD)
   return false;
@@ -213,31 +212,33 @@ void StarboardBridge::Initialize(JNIEnv* env, jobject obj) {
 
 int64_t StarboardBridge::GetAppStartTimestamp(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getAppStartTimestamp(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getAppStartTimestamp(env,
+                                                       j_starboard_bridge_);
 }
 
 void StarboardBridge::ApplicationStarted(JNIEnv* env) {
   SB_DCHECK(env);
-  Java_AOSPStarboardBridge_applicationStarted(env, j_starboard_bridge_);
+  Java_BaseStarboardBridge_applicationStarted(env, j_starboard_bridge_);
 }
 
 void StarboardBridge::ApplicationStopping(JNIEnv* env) {
   SB_DCHECK(env);
-  Java_AOSPStarboardBridge_applicationStopping(env, j_starboard_bridge_);
+  Java_BaseStarboardBridge_applicationStopping(env, j_starboard_bridge_);
 }
 
 void StarboardBridge::AppendArgs(JNIEnv* env,
                                  std::vector<std::string>* args_vector) {
   SB_DCHECK(env);
   ScopedJavaLocalRef<jobjectArray> args_java =
-      Java_AOSPStarboardBridge_getArgs(env, j_starboard_bridge_);
+      Java_BaseStarboardBridge_getArgs(env, j_starboard_bridge_);
   AppendJavaStringArrayToStringVector(env, args_java, args_vector);
 }
 
 ScopedJavaLocalRef<jintArray> StarboardBridge::GetSupportedHdrTypes(
     JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getSupportedHdrTypes(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getSupportedHdrTypes(env,
+                                                       j_starboard_bridge_);
 }
 
 void StarboardBridge::RaisePlatformError(JNIEnv* env,
@@ -245,127 +246,129 @@ void StarboardBridge::RaisePlatformError(JNIEnv* env,
                                          jlong data,
                                          const std::string& url) {
   SB_DCHECK(env);
-  Java_AOSPStarboardBridge_raisePlatformError(env, j_starboard_bridge_, errorType,
-                                          data,
-                                          ConvertUTF8ToJavaString(env, url));
+  Java_BaseStarboardBridge_raisePlatformError(
+      env, j_starboard_bridge_, errorType, data,
+      ConvertUTF8ToJavaString(env, url));
 }
 
 bool StarboardBridge::IsPlatformErrorShowing(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_isPlatformErrorShowing(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_isPlatformErrorShowing(env,
+                                                         j_starboard_bridge_);
 }
 
 void StarboardBridge::RequestSuspend(JNIEnv* env) {
   SB_DCHECK(env);
-  Java_AOSPStarboardBridge_requestSuspend(env, j_starboard_bridge_);
+  Java_BaseStarboardBridge_requestSuspend(env, j_starboard_bridge_);
 }
 
 ScopedJavaLocalRef<jobject> StarboardBridge::GetTextToSpeechHelper(
     JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getTextToSpeechHelper(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getTextToSpeechHelper(env,
+                                                        j_starboard_bridge_);
 }
 
 ScopedJavaLocalRef<jobject> StarboardBridge::GetCaptionSettings(JNIEnv* env) {
   SB_CHECK(env);
-  return Java_AOSPStarboardBridge_getCaptionSettings(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getCaptionSettings(env, j_starboard_bridge_);
 }
 
 ScopedJavaLocalRef<jobject> StarboardBridge::GetResourceOverlay(JNIEnv* env) {
   SB_CHECK(env);
-  return Java_AOSPStarboardBridge_getResourceOverlay(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getResourceOverlay(env, j_starboard_bridge_);
 }
 
 ScopedJavaLocalRef<jstring> StarboardBridge::GetSystemLocaleId(JNIEnv* env) {
   SB_CHECK(env);
-  return Java_AOSPStarboardBridge_systemGetLocaleId(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_systemGetLocaleId(env, j_starboard_bridge_);
 }
 
 SB_EXPORT_ANDROID std::string StarboardBridge::GetAdvertisingId(JNIEnv* env) {
   SB_DCHECK(env);
   ScopedJavaLocalRef<jstring> advertising_id_java =
-      Java_AOSPStarboardBridge_getAdvertisingId(env, j_starboard_bridge_);
+      Java_BaseStarboardBridge_getAdvertisingId(env, j_starboard_bridge_);
   return ConvertJavaStringToUTF8(env, advertising_id_java);
 }
 
 SB_EXPORT_ANDROID bool StarboardBridge::GetLimitAdTracking(JNIEnv* env) {
   SB_DCHECK(env);
   jboolean limit_ad_tracking_java =
-      Java_AOSPStarboardBridge_getLimitAdTracking(env, j_starboard_bridge_);
+      Java_BaseStarboardBridge_getLimitAdTracking(env, j_starboard_bridge_);
   return limit_ad_tracking_java == JNI_TRUE;
 }
 
 SB_EXPORT_ANDROID std::string StarboardBridge::GetFriendlyName(JNIEnv* env) {
   SB_DCHECK(env);
   ScopedJavaLocalRef<jstring> friendly_name_java =
-      Java_AOSPStarboardBridge_getFriendlyName(env, j_starboard_bridge_);
+      Java_BaseStarboardBridge_getFriendlyName(env, j_starboard_bridge_);
   return ConvertJavaStringToUTF8(env, friendly_name_java);
 }
 
 SB_EXPORT_ANDROID double StarboardBridge::GetScreenDiagonal(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getScreenDiagonal(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getScreenDiagonal(env, j_starboard_bridge_);
 }
 
 SB_EXPORT_ANDROID void StarboardBridge::CloseApp(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_closeApp(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_closeApp(env, j_starboard_bridge_);
 }
 
 std::string StarboardBridge::GetTimeZoneId(JNIEnv* env) {
   SB_DCHECK(env);
   ScopedJavaLocalRef<jstring> timezone_id_java =
-      Java_AOSPStarboardBridge_getTimeZoneId(env, j_starboard_bridge_);
+      Java_BaseStarboardBridge_getTimeZoneId(env, j_starboard_bridge_);
   return ConvertJavaStringToUTF8(env, timezone_id_java);
 }
 
 ScopedJavaLocalRef<jobject> StarboardBridge::GetDisplayDpi(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getDisplayDpi(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getDisplayDpi(env, j_starboard_bridge_);
 }
 
 Size StarboardBridge::GetDeviceResolution(JNIEnv* env) {
   SB_CHECK(env);
   ScopedJavaLocalRef<jobject> j_size =
-      Java_AOSPStarboardBridge_getDisplaySize(env, j_starboard_bridge_);
+      Java_BaseStarboardBridge_getDisplaySize(env, j_starboard_bridge_);
   return {Java_Size_getWidth(env, j_size), Java_Size_getHeight(env, j_size)};
 }
 
 void StarboardBridge::ReportFullyDrawn(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_reportFullyDrawn(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_reportFullyDrawn(env, j_starboard_bridge_);
 }
 
 void StarboardBridge::SetCrashContext(JNIEnv* env,
                                       const char* key,
                                       const char* value) {
   SB_CHECK(env);
-  Java_AOSPStarboardBridge_setCrashContext(env, j_starboard_bridge_,
-                                       ConvertUTF8ToJavaString(env, key),
-                                       ConvertUTF8ToJavaString(env, value));
+  Java_BaseStarboardBridge_setCrashContext(env, j_starboard_bridge_,
+                                           ConvertUTF8ToJavaString(env, key),
+                                           ConvertUTF8ToJavaString(env, value));
 }
 
 bool StarboardBridge::IsMicrophoneDisconnected(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_isMicrophoneDisconnected(env,
-                                                       j_starboard_bridge_);
+  return Java_BaseStarboardBridge_isMicrophoneDisconnected(env,
+                                                           j_starboard_bridge_);
 }
 
 bool StarboardBridge::IsMicrophoneMute(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_isMicrophoneMute(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_isMicrophoneMute(env, j_starboard_bridge_);
 }
 
 ScopedJavaLocalRef<jobject> StarboardBridge::GetAudioPermissionRequester(
     JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getAudioPermissionRequester(env,
-                                                          j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getAudioPermissionRequester(
+      env, j_starboard_bridge_);
 }
 
 void StarboardBridge::ResetVideoSurface(JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_resetVideoSurface(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_resetVideoSurface(env, j_starboard_bridge_);
 }
 
 void StarboardBridge::SetVideoSurfaceBounds(JNIEnv* env,
@@ -374,38 +377,41 @@ void StarboardBridge::SetVideoSurfaceBounds(JNIEnv* env,
                                             int width,
                                             int height) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_setVideoSurfaceBounds(env, j_starboard_bridge_, x,
-                                                    y, width, height);
+  return Java_BaseStarboardBridge_setVideoSurfaceBounds(
+      env, j_starboard_bridge_, x, y, width, height);
 }
 
 ScopedJavaLocalRef<jobject> StarboardBridge::GetAudioOutputManager(
     JNIEnv* env) {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getAudioOutputManager(env, j_starboard_bridge_);
+  return Java_BaseStarboardBridge_getAudioOutputManager(env,
+                                                        j_starboard_bridge_);
 }
 
 std::string StarboardBridge::GetUserAgentAuxField(JNIEnv* env) const {
   SB_DCHECK(env);
   return ConvertJavaStringToUTF8(
-      env, Java_AOSPStarboardBridge_getUserAgentAuxField(env, j_starboard_bridge_));
+      env,
+      Java_BaseStarboardBridge_getUserAgentAuxField(env, j_starboard_bridge_));
 }
 
 bool StarboardBridge::IsAmatiDevice(JNIEnv* env) const {
   SB_DCHECK(env);
-  return Java_AOSPStarboardBridge_getIsAmatiDevice(env, j_starboard_bridge_) ==
+  return Java_BaseStarboardBridge_getIsAmatiDevice(env, j_starboard_bridge_) ==
          JNI_TRUE;
 }
 
 std::string StarboardBridge::GetBuildFingerprint(JNIEnv* env) const {
   SB_DCHECK(env);
   return ConvertJavaStringToUTF8(
-      env, Java_AOSPStarboardBridge_getBuildFingerprint(env, j_starboard_bridge_));
+      env,
+      Java_BaseStarboardBridge_getBuildFingerprint(env, j_starboard_bridge_));
 }
 
 int64_t StarboardBridge::GetPlayServicesVersion(JNIEnv* env) const {
   SB_DCHECK(env);
-  return static_cast<int64_t>(
-      Java_AOSPStarboardBridge_getPlayServicesVersion(env, j_starboard_bridge_));
+  return static_cast<int64_t>(Java_BaseStarboardBridge_getPlayServicesVersion(
+      env, j_starboard_bridge_));
 }
 
 ScopedJavaLocalRef<jobject> StarboardBridge::OpenCobaltService(
@@ -413,7 +419,7 @@ ScopedJavaLocalRef<jobject> StarboardBridge::OpenCobaltService(
     jlong native_service,
     const char* service_name) {
   SB_CHECK(env);
-  return Java_AOSPStarboardBridge_openCobaltService(
+  return Java_BaseStarboardBridge_openCobaltService(
       env, j_starboard_bridge_, native_service,
       ConvertUTF8ToJavaString(env, service_name));
 }
@@ -421,35 +427,36 @@ ScopedJavaLocalRef<jobject> StarboardBridge::OpenCobaltService(
 void StarboardBridge::CloseCobaltService(JNIEnv* env,
                                          const char* service_name) {
   SB_CHECK(env);
-  Java_AOSPStarboardBridge_closeCobaltService(
+  Java_BaseStarboardBridge_closeCobaltService(
       env, j_starboard_bridge_, ConvertUTF8ToJavaString(env, service_name));
 }
 
 bool StarboardBridge::HasCobaltService(JNIEnv* env, const char* service_name) {
   SB_CHECK(env);
-  return Java_AOSPStarboardBridge_hasCobaltService(
+  return Java_BaseStarboardBridge_hasCobaltService(
       env, j_starboard_bridge_, ConvertUTF8ToJavaString(env, service_name));
 }
 
 void StarboardBridge::CloseAllCobaltService(JNIEnv* env) const {
   SB_DCHECK(env);
-  Java_AOSPStarboardBridge_closeAllCobaltService(env, j_starboard_bridge_);
+  Java_BaseStarboardBridge_closeAllCobaltService(env, j_starboard_bridge_);
 }
 
 void StarboardBridge::HideSplashScreen(JNIEnv* env) const {
   SB_DCHECK(env);
-  Java_AOSPStarboardBridge_hideSplashScreen(env, j_starboard_bridge_);
+  Java_BaseStarboardBridge_hideSplashScreen(env, j_starboard_bridge_);
 }
 
 void StarboardBridge::SetStartupMilestone(jint milestone) const {
   JNIEnv* env = AttachCurrentThread();
-  Java_AOSPStarboardBridge_setStartupMilestone(env, j_starboard_bridge_, milestone);
+  Java_BaseStarboardBridge_setStartupMilestone(env, j_starboard_bridge_,
+                                               milestone);
 }
 
 void StarboardBridge::SetStartupDiagnosisInfo(const char* key,
                                               const char* value) const {
   JNIEnv* env = AttachCurrentThread();
-  Java_AOSPStarboardBridge_setStartupDiagnosisInfo(
+  Java_BaseStarboardBridge_setStartupDiagnosisInfo(
       env, j_starboard_bridge_, ConvertUTF8ToJavaString(env, key),
       ConvertUTF8ToJavaString(env, value));
 }
